requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pytz==2023.3.post1
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
